# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Employee Asset management intended to use internally only. Not a public facing api.",
        "title": "EmployeeAssetManagement",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:4000",
    "basePath": "/",
    "paths": {
...
        "/asset/createasset": {
            "post": {
                "description": "Create a new Asset with the input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Create a new Asset",
                "parameters": [
                    {
                        "description": "Create Asset",
                        "name": "Asset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Asset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/asset/editasset/{id}": {
            "put": {
                "description": "updates an asset with the input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Asset"
                ],
                "summary": "Updates Asset",
                "parameters": [
                    {
                        "description": "Update asset",
                        "name": "asset",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Asset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/employee/createemployee": {
            "post": {
                "description": "Create a new employee with the input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Create a new employee",
                "parameters": [
                    {
                        "description": "Create Employee",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Employee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/employee/editemployee/{id}": {
            "put": {
                "description": "updates an employee with the input paylod",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Employee"
                ],
                "summary": "Updates Employee",
                "parameters": [
                    {
                        "description": "Update Employee",
                        "name": "employee",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Employee"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "model.Asset": {
            "type": "object",
            "properties": {
                "assetid": {
                    "type": "string"
                },
                "assetname": {
                    "type": "string",
                    "example": "Laptop"
                },
                "assettype": {
                    "type": "string",
                    "example": "Electronic"
                }
            }
        },
        "model.Employee": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "mumbai"
                },
                "bloodgroup": {
                    "type": "string",
                    "example": "AB-ve"
                },
                "emergencycontact": {
                    "type": "integer",
                    "example": 84984894
                },
                "empid": {
                    "type": "string"
                },
                "employeeemail": {
                    "type": "string",
                    "example": "ppp@gmail.com"
                },
                "firstname": {
                    "type": "string",
                    "example": "Shreyas"
                },
                "gender": {
                    "type": "string",
                    "example": "male"
                },
                "lastname": {
                    "type": "string",
                    "example": "Prabhu"
                },
                "phonenumber": {
                    "type": "integer",
                    "example": 8787878787
                }
            }
        },
...
    }
}